enable 64 bit pointers for GetHexString()

+++ source/library/common/utility.h
@@ -2,7 +2,7 @@
 #define UTILITY_H
 
 void HexDump (const void *, long, BString &, BString &, int32 BytesPerLine = 16);
-void GetHexString (char *, int32);
+void GetHexString (char *, intptr_t);
 
 #endif //UTILITY_H
